Improve some performance of is_static, is_dynamic - improve tensor check type by reduce number of validation
